Change usize to u8
At least where it makes sense so far.
It spares us lots of space.

+++ src/lib.rs
@@ -51,7 +51,7 @@ impl LeagueContract {
         league_name: &str,
         players: &[String],
         accounts: &[AccountId],
-        best_of: usize,
+        best_of: u8,
         game_type: GameType,
     ) {
         require!(best_of % 2 == 1, "best_of number should be odd");
